ONE-WIRE UP! desde debug y wifi, agrega menu Temp
El misterio del one-wire eran varias cosas, por un lado el int aca es de
32 y estaba todo para int de 16 asi que pase todo a unsigned int, pero
eso no fue todo.. tenia una operacion de tomar un arreglo de chars y
pasarlo a int usando biggle indian pero aca es little. asi que habia un
problema ahi tambien.. despues de eso salio todo de lujo.. Claro que
esta todo con delays de huevo y sin parar irqs. con lo cual si el uC se
chifla salta el CRC con error porque le pifia en la lectura.. tengo que
ver si lo dejo asi o uso irq;s.. lo que pasa es que los tiempos son
cortos.. del orden de los 100useg por cada bit...y mando comandos de 8
bits o  11 bits.. quien sabe partir todo en 1 bit con una maquinita...no
se... por ahora queda... veo si meto grabar parametros en un file en
flash para cerrar ese tema..

+++ h/one_wire_network.h
@@ -28,8 +28,8 @@ enum One_Wire_Network_Events
 typedef struct Rom_Codes_Struct
 {
  unsigned char Code[8];		//guarda el cogido del nodo...
- unsigned int T;		//guarda la temperatura actual del nodo, 
- unsigned int V;		//guarda la tension medida en el modelo DS2438
+ unsigned short int T;		//guarda la temperatura actual del nodo, 
+ unsigned short int V;		//guarda la tension medida en el modelo DS2438
  unsigned char Crc;		//guarda la tension medida en el modelo DS2438
 };
 //------------------------------------------------------ 
@@ -47,13 +47,13 @@ extern void 		Broadcast_V		(void);
 extern void 		Broadcast_T		(void);	
 extern void 		Broadcast_Recall_Page0	(void);
 //----------------- ALL -------------------------------
-extern unsigned char 	On_Line_Nodes		(void);
-extern unsigned char 	Family_Code		(unsigned char Node);
-extern void 		Print_Rom_Codes		(void);
-extern void 		Search_Codes		(void);
-extern unsigned int 	Convert_Nodes_Bin2Ascci	(unsigned char* Buf);
-extern unsigned char 	Read_One_Wire_Crc	(unsigned char Node);
-extern unsigned int 	Read_One_Wire_T		(unsigned char Node);
+extern unsigned char 		On_Line_Nodes		(void);
+extern unsigned char 		Family_Code		(unsigned char Node);
+extern void 			Print_Rom_Codes		(void);
+extern void 			Search_Codes		(void);
+extern unsigned short int 	Convert_Nodes_Bin2Ascci	(unsigned char* Buf);
+extern unsigned char 		Read_One_Wire_Crc	(unsigned char Node);
+extern unsigned short int 	Read_One_Wire_T		(unsigned char Node);
 //------------------DS18S20------------------------------------ 
 extern void 		Read_DS18S20_Scratchpad		(unsigned char Node);
 extern void 		Calculate_DS18S20_12Bit_T	(unsigned char Node);
